PR: 전시 상세 조회/즐겨찾기/리뷰 기능 구현 (#23)
* feat: ExhibitionDetailScreen에 exhibitId path/nav argument 추가

* feat: 마이페이지에 전시 상세 이동 임시 버튼 추가

* feat: 전시 상세 조회 UseCase 추가

* feat: 전시 상세 화면 앱바 뒤로가기 기능 추가

* refactor: ExhibitDetailModel을 ExhibitionDetailModel로 네이밍 변경

* feat: 전시 상세 조회 MVI 로직 추가

* feat: 전시 상세 화면에 실제 조회 데이터 바인딩

-  더미 데이터 제거 후 ExhibitionDetailState 값을 화면 UI에 바인딩하도록 변경
  - 포스터 이미지, 홈페이지 바로가기(웹 이동)
  - 전시 기간/전시명 표시
  - 상세 탭 정보(시간/전화번호/설명/주소) 노출
   - 주소 복사 기능 추가

* chore: lint 적용

* feat: 리뷰 조회 API 인터페이스 및 ResponseDto 추가

* feat: ExhibitRepository DI 바인딩 추가

* feat: 전시 상세 화면 앱바 뒤로가기 기능 추가

* feat: 전시 상세 화면에 실제 조회 데이터 바인딩

-  더미 데이터 제거 후 ExhibitionDetailState 값을 화면 UI에 바인딩하도록 변경
  - 포스터 이미지, 홈페이지 바로가기(웹 이동)
  - 전시 기간/전시명 표시
  - 상세 탭 정보(시간/전화번호/설명/주소) 노출
   - 주소 복사 기능 추가

* chore: lint 적용

* feat: 즐겨찾기(bookmark) MVI 상태 추가 및 screen에 적용

* feat: 전시상세의 즐겨찾기 토글 Optimistic UI 및 요청 직렬화 적용

즐겨찾기 토글은 optimistic update로 UI 즉시 반영하고, 서버 요청은 직렬화해서 동시에 1개만 보내도록 처리
-연타 시에는 마지막 의도로 수렴(pending target)

* feat: 즐겨찾기 API/DTO/매핑 모델 추가

* feat: 즐겨찾기 데이터 레이어 추가 및 API 호출 방식 통일

- 즐겨찾기 DataSource 및 구현 추가
- FavoriteRepository 인터페이스/구현체(FavoriteRepositoryImpl) 생성
- 즐겨찾기 mapper 및 매핑용 모델들 추가
- 즐겨찾기 API 호출 방식을 Call 기반에서 suspend 함수로 일괄 변경
- Hilt Module에 FavoriteRepositoryImpl 바인딩 추가

* refactor: domain 레이어 즐겨찾기 모델을 Bookmark 기준으로 정리

- FavoriteRepository → BookmarkRepository로 변경
- Favorite 도메인 모델을 Bookmark 도메인 모델로 정리
- data 레이어에서는 서버 용어(Favorite) 유지, domain/presentation은 앱 용어(Bookmark) 사용

* refactor: Favorite/Auth의 ReqDto/ResDto 네이밍 컨벤션으로 정리

* feat: Bookmark용 UseCase 추가

- add/remove/check

* feat: NetworkModule에 FavoriteApi Provider 추가

* refactor: exhibitId 타입을 Long에서 Int로 변경

* refactor: 도메인 레이어 모델 속성명 isFavorite → isBookmarked로 변경

* feat: 전시 상세 즐겨찾기 API 연동 및 토글 Optimistic 처리

- 즐겨찾기 토글 동작은 Optimistic Update로 UI 즉시 반영
- 연타 방지를 위해 debounce 적용 및 요청 중복 실행 방지 로직 추가

* feat: 전시 상세 탭 정보 없음 UI 처리 추가

- 전시 상세 탭에서 운영시간/주소/번호 값이 없을 경우 '정보없음'으로 표시
- 설명(description)이 비어있으면 해당 섹션 박스 자체를 숨기도록 처리

* fix: 리프레쉬 토큰 갱신 시 API 중복 호출 문제 해결

- synchronized 블록을 적용하여 토큰 갱신 로직의 스레드 안전성 확보

* style: 전시 상세 주소 텍스트 줄바꿈 및 복사 버튼 정렬 개선

* refactor: ReviewDetailPageResponseDto 컨벤션 네이밍 적용

* chore: Paging 라이브러리 의존성 추가

- androidx.paging runtime 의존성 추가
- androidx.paging compose 의존성 추가

* feat: 리뷰 조회 API 및 DTO/모델 추가

* feat: 리뷰 매핑 로직 및 DataSource 추가

* feat: 리뷰 Repository 추가 및 DI 구성

* feat: 리뷰 UseCase, PagingSource 추가

* feat: 전시 상세 리뷰 조회 MVI 연동

* refactor: 리뷰 탭을 LazyColumn 기반으로 전환 및 실제 데이터 바인딩

- 페이징 적용을 위해 VerticalScroll 기반 레이아웃을 LazyColumn으로 변경
- 하위 아이템들을 LazyListScope 확장 함수 형태로 분리/정리
- 리뷰 탭의 더미 데이터 제거
- 리뷰 목록을 실제 조회 데이터로 바인딩하도록 수정
-Tabsection 제거

+++ app/src/main/java/com/arttrip/android/data/remote/model/home/ExhibitDetailResponseDto.kt
@@ -1,7 +1,7 @@
 package com.arttrip.android.data.remote.model.home
 
 data class ExhibitDetailResponseDto(
-    val exhibitId: Long,
+    val exhibitId: Int,
     val title: String,
     val description: String,
     val posterUrl: String?,
